refactor(inference): move detokenize into tokenizer module (#1302)
Moves model/qwen35/detokenize.rs to tokenizer/detokenize.rs, breaking the
tokenizer -> model::qwen35 -> tokenizer module cycle. Updates all call
sites (metal_qwen35.rs, generation.rs, qwen35/mod.rs, stop_strings.rs) to
import from the new location.

Widens BpeTokenizer::from_vocab_and_merges_with_config and the
DEFAULT_BPE_CACHE_CAPACITY/DEFAULT_BPE_MAX_SEQ_LEN constants to
pub(crate) so detokenize.rs tests (moved from bpe.rs, where they tested
detokenize behavior, not bpe behavior) can construct tokenizers directly.
